muax/requirements.txt
trfl
optax
gym[atari]
dm-haiku==0.0.10
dm-reverb==0.7.2
dm-control
numpy
atari-py
dm-env
dm-tree
tensorflow_probability==0.15.0
dm-acme
rlds
optax==0.1.7
tensorflow==2.8.0
gym==0.25.0
bsuite
pygame==2.1.0
tensorboardX
flax
mctx
pillow
dm-launchpad==0.5.2
gymnasium
dm-sonnet
lz4
rlax==0.1.6
chex==0.1.6
tensorflow_datasets==4.6.0
jax==0.4.3
dm-haiku
jaxlib==0.4.3
absl-py
typing-extensions
